+++ my-profile-app/src/components/NavbarComponent/NavbarComponent.module.css
@@ -1,0 +1,117 @@
+/* Navbar Styles */
+.navbar {
+    background-color: var(--background-color);
+    padding: 1rem 2rem;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    position: relative;
+}
+
+.navbarContainer {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+}
+
+.navbarBrand {
+    font-size: 1.125rem;
+    font-weight: bold;
+    color: var(--primary-color);
+    text-decoration: none;
+}
+
+.navbarNav {
+    display: flex;
+    list-style: none;
+    gap: 1.5rem;
+    margin: 0;
+    padding: 0;
+    align-items: center;
+    flex-grow: 1; /* Pushes the Resume button to the far right */
+}
+
+.navItem {
+    list-style: none;
+}
+
+.navLink {
+    color: var(--secondary-text-color);
+    text-decoration: none;
+    font-size: 1rem;
+    transition: color 0.3s ease;
+}
+
+.navLink:hover {
+    color: var(--primary-color);
+}
+
+.resumeButton {
+    padding: 0.5rem 1rem;
+    border: 1px solid var(--primary-color);
+    color: var(--primary-color);
+    text-decoration: none;
+    border-radius: 4px;
+    font-size: 1rem;
+    transition: background-color 0.3s ease, color 0.3s ease;
+    margin-left: auto; /* Pushes the Resume button to the far right */
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.resumeButton:hover {
+    background-color: var(--primary-color);
+    color: var(--background-color);
+}
+
+.navbarToggler {
+    display: none;
+    background: none;
+    border: none;
+    color: var(--primary-color);
+    font-size: 1.5rem;
+    cursor: pointer;
+    align-self: center;
+}
+
+.navbarTogglerIcon {
+    display: block;
+    width: 1.5rem;
+    height: 0.125rem;
+    background-color: var(--primary-color);
+    margin: 0.25rem 0;
+}
+
+
+/* Responsive Navbar */
+@media (max-width: 768px) {
+    .navbarToggler {
+        display: block;
+    }
+
+    .navbarCollapse {
+        display: none;
+        flex-direction: column;
+        align-items: flex-start;
+        background-color: var(--background-color);
+        padding: 1rem;
+        position: absolute;
+        top: 100%;
+        left: 0;
+        width: 100%;
+        z-index: 10;
+    }
+
+    .navbarCollapse.show {
+        display: flex;
+    }
+
+    .resumeButton {
+        margin-left: 0;
+        margin-top: 1rem;
+        width: 100%;
+        text-align: center;
+    }
+}
